# LaravelProject2/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "summary": "register",
                "description": "register the user for login",
                "operationId": "2f25885f3b6f61ca11724ec302d2e439",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "firstname",
                                    "lastname",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "firstname": {
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User successfully registered"
                    },
                    "401": {
                        "description": "The email has already been taken"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "summary": "login",
                "description": "login",
                "operationId": "0fc9610184083ac6dddab2c63139225a",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "login Success"
                    },
                    "401": {
                        "description": "we can not find the user with that e-mail address You need to register first"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "summary": "logout",
                "description": "logout",
                "operationId": "67d5147477c548bfbc3f982833bbc7bd",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "token"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User successfully registered"
                    },
                    "401": {
                        "description": "The email has already been taken"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/getuser": {
            "get": {
                "summary": "get user",
                "description": "get user",
                "operationId": "7508d38362951e3be4c31513644ab5e4",
                "requestBody": {},
                "responses": {
                    "201": {
                        "description": "Found User successfully"
                    },
                    "401": {
                        "description": "User cannot be found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/forgotPassword": {
            "post": {
                "summary": "forgot password",
                "description": "forgot user password",
                "operationId": "e6aaf1796b83d83a4874db2d170e0199",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password Reset link is send to your email"
                    },
                    "400": {
                        "description": "we can not find a user with that email address"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/resetPassword": {
            "post": {
                "summary": "reset password",
                "description": "reset user password",
                "operationId": "61516897d7ad1e16ea9a0dc0dadede33",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "new_password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "new_password": {
                                        "type": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset successfull!"
                    },
                    "400": {
                        "description": "we can't find the user with that e-mail address"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "laravel bearer authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}